+++ MyMoney.Web/FrontEnd/apps/mymoney-react/src/app/state/budgets-list/constants.ts
@@ -1,0 +1,1 @@
+export const SLICE_NAME = 'budgetsList';
