+++ MyMoney.Web/FrontEnd/apps/mymoney-react/src/app/state/budgets-list/slice.ts
@@ -1,11 +1,8 @@
-import { createSlice, createAsyncThunk, ActionReducerMapBuilder } from '@reduxjs/toolkit';
-import { BudgetApi, IBudgetDto, IBudgetListDto, IBudgetSearchDto } from '@mymoney-common/api';
+import { createSlice, ActionReducerMapBuilder } from '@reduxjs/toolkit';
+import { IBudgetDto, IBudgetSearchDto } from '@mymoney-common/api';
 import { AsyncStatus, IBudgetState } from '../types';
-import { map } from 'rxjs/operators';
-import { firstValueFrom } from 'rxjs';
-import { HttpHelper } from '../../classess/http-helper';
-
-const SLICE_NAME = 'budgetsList';
+import { SLICE_NAME } from './constants';
+import { deleteBudget, fetchBudgets } from './thunks';
 
 export const initialBudgetsState: IBudgetState = {
    budgets: {
@@ -19,40 +16,6 @@ export const initialBudgetsState: IBudgetState = {
       refresh: true,
    },
 };
-
-export const fetchBudgets = createAsyncThunk(
-   `${SLICE_NAME}/fetchBudgets`,
-   async ({ search }: { search: IBudgetSearchDto }, { getState, rejectWithValue }) => {
-      const httpHelper = HttpHelper.forCuurentUser(getState);
-      if (!httpHelper) {
-         return rejectWithValue('No user session');
-      }
-      const api = new BudgetApi(httpHelper);
-
-      try {
-         return await firstValueFrom(api.list(search).pipe(map((listDto: IBudgetListDto) => listDto.budgets)));
-      } catch (error: any) {
-         return rejectWithValue(error.message);
-      }
-   }
-);
-
-export const deleteBudget = createAsyncThunk(
-   `${SLICE_NAME}/deleteBudget`,
-   async ({ budgetId }: { budgetId: number }, { getState, rejectWithValue }) => {
-      const httpHelper = HttpHelper.forCuurentUser(getState);
-      if (!httpHelper) {
-         return rejectWithValue('No user session');
-      }
-      const api = new BudgetApi(httpHelper);
-
-      try {
-         return await firstValueFrom(api.delete({ id: budgetId }));
-      } catch (error: any) {
-         return rejectWithValue(error.message);
-      }
-   }
-);
 
 export const budgetsSlice = createSlice({
    name: SLICE_NAME,
