+++ MyMoney.Web/FrontEnd/apps/mymoney-react/src/app/state/running-total-chart/constants.ts
@@ -1,0 +1,1 @@
+export const SLICE_NAME = 'runningTotalChart';
