+++ MyMoney.Web/FrontEnd/apps/mymoney-react/src/app/state/transactions-list/index.ts
@@ -1,2 +1,3 @@
 export * from './selectors';
 export * from './slice';
+export * from './thunks';
